+++ src/Config.php
@@ -40,17 +40,16 @@ class Config
      */
     private $input;
 
-    const MISSING_SHA_ERROR = "Please ensure that the GITHUB_SHA or CIRCLE_SHA1 env variable contains the commit SHA.";
-    const MISSING_TOKEN_ERROR = "Please ensure that the APPOCULAR_TOKEN env variable contains the token for Appocular.";
-    const MISSING_WEBDRIVER_URL = 'Please provide a webdriver url, either in the config file or with the --webdriver option.';
-    const MISSING_BASE_URL = 'Please provide a base url, either in the config file or with the --base-url option.';
+    public const MISSING_SHA_ERROR = "Please ensure that the GITHUB_SHA or CIRCLE_SHA1 env variable contains the commit SHA.";
+    public const MISSING_TOKEN_ERROR = "Please ensure that the APPOCULAR_TOKEN env variable contains the token for Appocular.";
+    public const MISSING_WEBDRIVER_URL = 'Please provide a webdriver url, either in the config file or with the --webdriver option.';
+    public const MISSING_BASE_URL = 'Please provide a base url, either in the config file or with the --base-url option.';
 
     public function __construct(Env $env, ConfigFile $configFile, InputInterface $input, Git $git)
     {
         try {
             $this->sha = $env->get('GITHUB_SHA', self::MISSING_SHA_ERROR);
-        }
-        catch (\RuntimeException $e) {
+        } catch (\RuntimeException $e) {
             $this->sha = $env->get('CIRCLE_SHA1', self::MISSING_SHA_ERROR);
         }
         if (empty($this->sha)) {
@@ -146,5 +145,11 @@ class Config
     public function getBase()
     {
         return $this->input->getOption('base') ? $this->input->getOption('base') : 'alpha.appocular.io';
+    }
+
+    public function getWorkers()
+    {
+        // Default to four workers.
+        return $this->configFile->getWorkers() ?? 4;
     }
 }