+++ src/Helpers/Output.php
@@ -26,6 +26,23 @@ class Output
         $formatter->setStyle('warning', new OutputFormatterStyle('yellow'));
         $formatter->setStyle('error', new OutputFormatterStyle('red'));
         $formatter->setStyle('timestamp', new OutputFormatterStyle('white'));
+
+        $colors = [
+            'blue',
+            'magenta',
+            'cyan',
+            'default',
+            'red',
+            'green',
+            'yellow',
+        ];
+
+        $i = 1;
+        foreach ($colors as $color) {
+            $formatter->setStyle('color' . $i, new OutputFormatterStyle($colors[$i - 1]));
+            $formatter->setStyle('color' . ($i + count($colors)), new OutputFormatterStyle($colors[$i - 1], null, ['reverse']));
+            $i++;
+        }
     }
 
     public function debug($message)
@@ -58,13 +75,20 @@ class Output
         $this->output->writeln('', OutputInterface::VERBOSITY_NORMAL);
     }
 
+    public function numberedLine($num, $message)
+    {
+        $colorIndex = (($num - 1) % 14 ) + 1;
+        $message = sprintf('<color%d>%d</> > %s', $colorIndex, $colorIndex, $message);
+        $this->output->writeln($message, OutputInterface::VERBOSITY_NORMAL);
+    }
+
     protected function writeln($style, $message, $verbosity, $errorOutput = false)
     {
         $message = sprintf("<%s>%s</>", $style, $message);
 
         // Add timestamps in debug mode.
         if ($this->output->isDebug()) {
-            $message = sprintf("[<timestamp>%s</>] %s", Carbon::now()->format('Y-m-d H:i:s'), $message);
+            $message = sprintf("<timestamp>[</>%s<timestamp>]</> %s", Carbon::now()->format('Y-m-d H:i:s'), $message);
         }
 
         // Print errors to stderr.